remove: Height 클래스 제거
